An-exercise-with-PIR-and-HC-SR04-sensors
This study implemented motion detection using a PIR (Passive Infrared) sensor and activated an alert light on an RGB LED when the distance measured by the HC-SR04 ultrasonic sensor fell between 0 and 50 cm. The ESP32 development board was utilized for this project, which was conducted in the Wokwi simulation environment.